requirements.txt
gradio>=4.19.2
huggingface-hub>=0.20.3
python-dotenv>=1.0.0 